Fixed const size to fit port values > 127

+++ 2017.03.18/mqtt_temp_sensor_esp8266/connection_settings.h
@@ -3,4 +3,4 @@
 const char*  ssid = ".......";
 const char*  password = ".......";
 const char*  mqtt_server = ".......";
-const int8_t mqtt_port = 1883;
+const int16_t mqtt_port = 1883;
